requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
ffmpeg-python==0.2.0
aiohttp==3.8.5
gunicorn==20.1.0
flask==2.2.5
numpy==1.24.3
tqdm==4.65.0
torch==2.1.0
openai-whisper==20231117